cleaning up code and finishing crew ai enedpoint
