+++ backend/app/models/email_summary.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class Email_Summary_Request(BaseModel):
+    content: str
