+++ backend/app/routers/company_job.py
@@ -10,6 +10,9 @@ import os
 import json
 from typing import Dict, List
 import re 
+from models.email_summary import Email_Summary_Request
+from models.create_table import JobInformation, ColumnResult
+
 router = APIRouter()
 load_dotenv()
 PERPLEXITY_API_KEY = os.getenv("PERPLEXITY_API_KEY")
@@ -20,23 +23,19 @@ ENDPOINT_OPENAI_GPT4 = os.getenv("AZURE_API_BASE")
 CHAT_VERSION = "2024-08-01-preview"  # Update if needed
 CHAT_DEPLOYMENT_NAME = "gpt-4o"  # Replace with your deployed model name
 
-"""
-perplexity_client = PerplexityClient(
-    api_key=PERPLEXITY_API_KEY,
-    api_url=PERPLEXITY_API_URL,
-    model="sonar"
-)
-"""
-print(PERPLEXITY_API_URL)
+def string_to_json(response):
+    cleaned = re.sub(r"^```json\s*", "", response)
+    cleaned = re.sub(r"\s*```$", "", cleaned)
+    return cleaned
+
 client = OpenAI(api_key=PERPLEXITY_API_KEY, base_url="https://api.perplexity.ai")
 
 
 
-class Email_Summary_Request(BaseModel):
-    content: str
+
 
 @router.post("/email-summary")
-async def get_summary_email(packet: Email_Summary_Request):
+async def get_summary_email(packet: modeEmail_Summary_Request):
     messages = [
             {
                 "role": "system", 
@@ -61,14 +60,7 @@ async def get_summary_email(packet: Email_Summary_Request):
         chat_client.add_memory(role = "assistant", content = i)
     return {"data": chat_client.get_history()[-1]}
 
-class ColumnResult(BaseModel):
-    status: str
-    content: str
-    source: List[str]
 
-class JobInformation(BaseModel):
-    company: str
-    results: Dict[str, ColumnResult]
 
 @router.get("/company-job-info")
 async def get_company_job_info(company: str, job_position: str):
@@ -127,12 +119,9 @@ async def get_company_job_info(company: str, job_position: str):
                                        api_url="https://api.perplexity.ai/chat/completions", 
                                        model="sonar")
     response = await prelixty_client.get_response(messages=messages, response_format=response_format)
-    print(type(response))
-    print(("Raw response from API: %r", response))
-    cleaned = re.sub(r"^```json\s*", "", response)
-    cleaned = re.sub(r"\s*```$", "", cleaned)
+    
     try:
-        response = json.loads(cleaned)
+        response = json.loads(string_to_json(response))
     except Exception as e:
         print("ERROR RESPONSE", e)
         raise HTTPException(status_code=500, detail="Response validation failed")
@@ -152,4 +141,17 @@ async def get_company_job_info(company: str, job_position: str):
         'company': company,
         'job': job_position
     }
-    LatestAiDevelopmentCrew().crew().kickoff(inputs=inputs)   
+    result = LatestAiDevelopmentCrew().crew().kickoff(inputs=inputs)   
+    try:
+        response = json.loads(string_to_json(result))
+    except Exception as e:
+        print("ERROR RESPONSE", e)
+        raise HTTPException(status_code=500, detail="Response validation failed")
+
+    try:
+        response_format = JobInformation(**response)
+    except Exception as e:
+        print("ERROR RESPONSE", e)
+        raise HTTPException(status_code=500, detail="Response validation failed")
+
+    return {"data": response_format}
